change app router on about page
